Write out the casts clang refuses
Three conversions that gcc performs with a warning and clang rejects outright,
all of them in code only the 32-bit x86 backend runs but that every target
compiles: a pointer truncated to int32 in EmitMEM, the same to uint32 in
PatchManager, and a function pointer passed to the FlushInstructionCache shim,
which takes void*.

Going through uintptr_t says what the 32-bit emitter means — the absolute
address is emitted as a 32-bit displacement, so the truncation is the point —
and leaves behaviour on x86 exactly as it was. What changes is that the tree now
compiles with a compiler other than gcc, which is what a macOS build needs.

+++ src/PatchManager.cpp
@@ -20,7 +20,11 @@ void PatchManager::ApplyPatches()
         //*((uint64 *)(patchData[i].patchPtr)) = (uint64)(labelPointers[patchData[i].destLabel] - patchData[i].basePtr);
         break;
       case PatchType::PatchType_Abs32:
-        *((uint32 *)(patchData[i].patchPtr)) = (uint32)(labelPointers[patchData[i].destLabel]); //!! 64bit prob?
+        // Absolute 32-bit patch slot: the pointer is truncated by design (and
+        // only ever emitted by the 32-bit x86 backend). Written through
+        // uintptr_t because clang rejects the direct pointer-to-uint32 cast
+        // that gcc only warns about.
+        *((uint32 *)(patchData[i].patchPtr)) = (uint32)(uintptr_t)(labelPointers[patchData[i].destLabel]); //!! 64bit prob?
         break;
       case PatchType::PatchType_Abs64:
         //*((uint64 *)(patchData[i].patchPtr)) = (uint32)(labelPointers[patchData[i].destLabel]);
